<BUG> To send data to the AstronomicalBodyPage, add a query parameter string to the routing information. The string is of the form ?astroName=astroBodyToDisplay.
To receive the data on the AstronomicalBodyPage, first create a class-level property to hold the incoming data. Name it AstroName.

Here, the UpdateAstroBodyUI(astroName) is a helper function used to update the user interface immediately when the AstroName property is set.

Then you need to decorate the class with an annotation that maps the incoming query parameter to the property you created.